requirements.txt

langchain>=0.1.16
langgraph
requests>=2.31.0
python-dotenv>=1.0.1
flask>=2.3.3
groq>=0.5.0
langchain-groq
